Added our panzoom script
