+++ docs/zensical.toml
@@ -12,7 +12,7 @@ extra_css = [
 ]
 extra_javascript = [
   "javascripts/shiki.mjs",
-  { path = "assets/javascripts/mermaid-panzoom.js" },
+  { path = "assets/javascripts/mermaid-panzoom.js", type = "module" },
 ]
 
 nav = [
